+++ app/(app)/mi-cuenta/page.tsx
@@ -1,5 +1,5 @@
 import { redirect } from "next/navigation";
 
-export default function MarketplacePayoutsAlias() {
+export default function AccountIndexPage() {
     redirect("/mi-cuenta/billetera");
 }
